+++ android/domain/src/main/java/com/now/domain/model/Adventure.kt
@@ -1,7 +1,14 @@
 package com.now.domain.model
 
+import java.time.LocalDateTime
+
 data class Adventure(
     val id: Long,
-    val destination: Destination,
+    val createdAt: LocalDateTime,
+    val startCoordinate: Coordinate,
+    val destination: Place,
+    val player: Player,
     val adventureStatus: AdventureStatus,
+    val remainingTryCount: RemainingTryCount,
+    val hints: List<Hint>,
 )
