--- android/domain/src/main/java/com/now/domain/model/Game.kt
@@ -1,14 +1,0 @@
-package com.now.domain.model
-
-import java.time.LocalDateTime
-
-data class Game(
-    val id: Long,
-    val createdAt: LocalDateTime,
-    val startCoordinate: Coordinate,
-    val destination: Place,
-    val player: Player,
-    val adventureStatus: AdventureStatus,
-    val remainingTryCount: RemainingTryCount,
-    val hints: List<Hint>,
-)
